bret comments
