+++ openmdao/recorders/recording_manager.py
@@ -3,7 +3,8 @@ RecordingManager class definition.
 """
 import time
 
-from openmdao.utils.general_utils import simple_warning, warn_deprecation
+from openmdao.utils.general_utils import simple_warning
+from openmdao.warnings import warn_deprecation
 
 try:
     from openmdao.utils.mpi import MPI
