+++ openmdao/utils/om.py
@@ -48,7 +48,8 @@ from openmdao.utils.entry_points import _list_installed_setup_parser, _list_inst
     split_ep, _compute_entry_points_setup_parser, _compute_entry_points_exec, \
         _find_plugins_setup_parser, _find_plugins_exec
 from openmdao.core.component import Component
-from openmdao.utils.general_utils import ignore_errors, warn_deprecation
+from openmdao.utils.general_utils import ignore_errors
+from openmdao.warnings import warn_deprecation
 
 
 def _n2_setup_parser(parser):
